+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # DailyCats
 
+## Release v1.2
+\- More verbose download script
+\- Change source too `thecatapi.com`
+
 ## Release v1.1
 \- Only download images tagged as `cute`